Break up venuedetailt in to smaller components
